+++ scripts/utils.ts
@@ -1,7 +1,7 @@
 export const files = {
   searchParams: {
     module: './src/config/search-params.ts',
-    importDir: ({ fileName }) => `../websites-services/${fileName}`,
+    importDir: ({ fileName }) => `../services/websites/${fileName}`,
     schema: './src/config/search-params.schema.json',
   },
 }
